refactor(full library): update to angular2-rc
BodyOutputType.Component (the ability to render components as the body
of the toast) has been updated to remove the soon to be deprecated
DynamicComponentLoader in favor of the ComponentResolver.

Toast rendering has been moved to its own component.

ToasterContainerComponent.Template The template has been updated to the
new *ngFor syntax.

Closes [#14](https://github.com/Stabzs/Angular2-Toaster/issues/14).