fix(toaster.css): toast icon no longer covers entire toast body
The toast icons were covering the entire toast body due to their layout.
As a result, any component being rendered into the toast body could not
have any interaction points (such as buttons) since it was beneath the
icon.  The toast has been slightly restructured to accommodate this
scenario.

An additional use case has been added to the systemjs demo to
demonstrate how to use a component with actionable items as the body.