feat(BodyOutputType.Component): added toast instance
When rendering BodyOutputType.Component, the toast instance itself is
applied to the rendered component.  This allows the component to
interact with the toast as needed, as well as expose access to the toast
id.  Addresses https://github.com/Stabzs/Angular2-Toaster/issues/84.

ResetTimer on mouseout was not being properly called. It is now called
appropriately when mouseoverTimerStop is set to true.  Thanks to @kb3eua
for the pull request fix.

Documented the mouseoverTimerStop config option and documented the
addition of the toast instance to components when using
BodyOutputType.Component.  Closes
https://github.com/Stabzs/Angular2-Toaster/issues/83.